% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{get_species}
\alias{get_species}
\title{return the species column for palmerpenguins}
\usage{
get_species()
}
\value{
dataframe of one column
}
\description{
return the species column for palmerpenguins
}
\examples{
get_species()

\dontrun{
get_species("hi")
}

}
